update apollo and rename package
