+++ android/Omnivore/app/src/main/java/app/omnivore/omnivore/AppModule.kt
@@ -2,7 +2,7 @@ package app.omnivore.omnivore
 
 import android.content.Context
 import app.omnivore.omnivore.dataService.DataService
-import app.omnivore.omnivore.networking.Networker
+import app.omnivore.omnivore.network.Networker
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
