+++ android/Omnivore/app/src/main/java/app/omnivore/omnivore/dataService/SavedItemLabelSync.kt
@@ -1,6 +1,6 @@
 package app.omnivore.omnivore.dataService
 
-import app.omnivore.omnivore.networking.savedItemLabels
+import app.omnivore.omnivore.network.savedItemLabels
 
 suspend fun DataService.syncLabels() {
   val fetchedLabels = networker.savedItemLabels()
