+++ android/Omnivore/app/src/main/java/app/omnivore/omnivore/network/HighlightMutations.kt
@@ -1,4 +1,4 @@
-package app.omnivore.omnivore.networking
+package app.omnivore.omnivore.network
 
 import android.util.Log
 import app.omnivore.omnivore.graphql.generated.CreateHighlightMutation
