+++ android/Omnivore/app/src/main/java/app/omnivore/omnivore/network/SavedItemLabelMutations.kt
@@ -1,4 +1,4 @@
-package app.omnivore.omnivore.networking
+package app.omnivore.omnivore.network
 
 import app.omnivore.omnivore.graphql.generated.CreateLabelMutation
 import app.omnivore.omnivore.graphql.generated.SetLabelsMutation
